Hooks: Only add NSS subdir path for QtWebEngine if it exists.
On Archlinux, the NSS package ships with /usr/lib/libnss3.so.

The auto detection works fine with that, but the fix from #3439 breaks it again
since no /usr/lib/nss/libnss3.so exists:

    INFO: Loading module hook "hook-PyQt5.QtWebEngineWidgets.py"...
    Unable to find "/usr/lib/nss/*.so" when adding binary and data files.

Thus, it should only be added on distributions where it's really necessary.
With this change, test_PyQt5_QWebEngine works on Archlinux (it failed before).